refactor: split binding source codegen
