+++ .sampo/changesets/1026-split-binding-source-codegen.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/block-types: patch
+---
+
+Split block binding source PHP and editor registration source generation into a
+focused internal codegen module while preserving the existing bindings facade.
